updated test_output_engine and refactored OE
