+++ cve_bin_tool/output_engine/console.py
@@ -1,0 +1,85 @@
+import textwrap
+
+from collections import defaultdict
+from datetime import datetime
+from typing import List, Dict
+
+from rich.console import Console
+from rich.markdown import Markdown
+from rich.panel import Panel
+from rich.table import Table
+
+from ..input_engine import Remarks
+
+
+def output_console(all_cve_data, console=Console(legacy_windows=False)):
+    """ Output list of CVEs in a tabular format with color support """
+
+    now = datetime.now().strftime("%Y-%m-%d  %H:%M:%S")
+
+    console.print(
+        Markdown(
+            textwrap.dedent(
+                f"""
+                # CVE BINARY TOOL
+                - cve-bin-tool Report Generated: {now}
+                """
+            )
+        )
+    )
+
+    # colors provide color name according to the severity
+    colors = {
+        "CRITICAL": "red",
+        "HIGH": "blue",
+        "MEDIUM": "yellow",
+        "LOW": "green",
+        "UNKNOWN": "white",
+    }
+
+    remarks_colors = {
+        Remarks.Mitigated: "green",
+        Remarks.Confirmed: "red",
+        Remarks.NewFound: "blue",
+        Remarks.Unexplored: "yellow",
+        Remarks.Ignored: "white",
+    }
+
+    cve_by_remarks: defaultdict[Remarks, List[Dict[str, str]]] = defaultdict(list)
+    # group cve_data by its remarks
+    for product in all_cve_data:
+        for cve in product.cves:
+            cve_by_remarks[product.remarks].append(
+                {
+                    "vendor": product.vendor,
+                    "product": product.product,
+                    "version": product.version,
+                    "cve_number": cve.cve_number,
+                    "severity": cve.severity,
+                }
+            )
+
+    for remarks in sorted(cve_by_remarks):
+        color = remarks_colors[remarks]
+        console.print(Panel(f"[{color}] {remarks.name} CVEs [/{color}]", expand=False))
+        # table instance
+        table = Table()
+
+        # Add Head Columns to the Table
+        table.add_column("Vendor")
+        table.add_column("Product")
+        table.add_column("Version")
+        table.add_column("CVE Number")
+        table.add_column("Severity")
+
+        for cve_data in cve_by_remarks[remarks]:
+            color = colors[cve_data["severity"]]
+            table.add_row(
+                f'[{color}]{cve_data["vendor"]} [/{color}]',
+                f'[{color}]{cve_data["product"]} [/{color}]',
+                f'[{color}]{cve_data["version"]} [/{color}]',
+                f'[{color}]{cve_data["cve_number"]} [/{color}]',
+                f'[{color}]{cve_data["severity"]} [/{color}]',
+            )
+        # Print the table to the console
+        console.print(table)
