+++ cve_bin_tool/output_engine/util.py
@@ -1,0 +1,76 @@
+"""
+output_engine/util.py - Provides helper functions for OutputEngine.
+"""
+
+import os
+
+from datetime import datetime
+from typing import List, Dict
+
+
+def generate_filename(self, extension) -> str:
+    """[summary]
+
+    Args:
+        extension str : Can be any value from list[CSV, JSON, HTML]  
+
+    Returns:
+        str: unique generated filename
+    """
+    now = datetime.now().strftime("%Y-%m-%d.%H-%M-%S")
+    filename = os.path.abspath(
+        os.path.join(os.getcwd(), f"output.cve-bin-tool.{now}.{extension}")
+    )
+    return filename
+
+
+def format_output(all_cve_data) -> List[Dict[str, str]]:
+    """
+    summary: format output in the list of dictionary format.
+
+    Returns:
+        formatted_output: List[Dict[str, str]]
+        - example:  [
+                        {
+                            "vendor": "haxx"
+                            "product": "curl",
+                            "version": "1.2.1",
+                            "cve_number": "CVE-1234-1234",
+                            "severity": "LOW"
+                        },
+                        ...
+                    ]
+    """
+    formatted_output = []
+    for product in all_cve_data:
+        for cve in product.cves:
+            formatted_output.append(
+                {
+                    "vendor": product.vendor,
+                    "product": product.product,
+                    "version": product.version,
+                    "remarks": product.remarks.name,
+                    "cve_number": cve.cve_number,
+                    "severity": cve.severity,
+                }
+            )
+    return formatted_output
+
+
+def add_extension_if_not(filename: str, output_type: str) -> str:
+    """
+    summary: Checks if the filename ends with the extension and if not 
+    adds one.
+
+    Args:
+        filename (str): filename from OutputEngine
+        output_type (str): contains a value from ["json", "csv", "html"]
+
+    Returns:
+        str: [description]
+    """
+    if not filename.endswith(f".{output_type}"):
+        updated_filename = f"{filename}.{output_type}"
+        return updated_filename
+    else:
+        return filename
