+++ test/test_output_engine.py
@@ -11,6 +11,10 @@ import unittest
 from rich.console import Console
 
 from cve_bin_tool.output_engine import OutputEngine
+from cve_bin_tool.output_engine import output_csv
+from cve_bin_tool.output_engine import output_json
+from cve_bin_tool.output_engine.console import output_console
+from cve_bin_tool.output_engine.util import format_output
 from cve_bin_tool.cve_scanner import CVE, CVEData
 from cve_bin_tool.input_engine import Remarks
 
@@ -86,17 +90,17 @@ class TestOutputEngine(unittest.TestCase):
 
     def test_formatted_output(self):
         """ Test reformatting products """
-        self.assertEqual(self.output_engine.format_output(), self.FORMATTED_OUTPUT)
+        self.assertEqual(format_output(self.MOCK_OUTPUT), self.FORMATTED_OUTPUT)
 
     def test_output_json(self):
         """ Test formatting output as JSON """
-        self.output_engine.output_json(self.mock_file)
+        output_json(self.MOCK_OUTPUT, self.mock_file)
         self.mock_file.seek(0)  # reset file position
         self.assertEqual(json.load(self.mock_file), self.FORMATTED_OUTPUT)
 
     def test_output_csv(self):
         """ Test formatting output as CSV """
-        self.output_engine.output_csv(self.mock_file)
+        output_csv(self.MOCK_OUTPUT, self.mock_file)
         self.mock_file.seek(0)  # reset file position
         reader = csv.DictReader(self.mock_file)
         self.assertEqual(list(reader), self.FORMATTED_OUTPUT)
@@ -105,7 +109,7 @@ class TestOutputEngine(unittest.TestCase):
         """Test Formatting Output as console"""
 
         console = Console(file=self.mock_file)
-        self.output_engine.output_console(console=console)
+        output_console(self.MOCK_OUTPUT, console=console)
         expected_output = "│ vendorname0  │ productname0  │ 1.0      │ CVE-1234-1234  │ MEDIUM   │\n│ vendorname0  │ productname0  │ 1.0      │ CVE-1234-9876  │ LOW      │\n│ vendorname0  │ productname0  │ 2.8.6    │ CVE-1234-1111  │ LOW      │\n│ vendorname1  │ productname1  │ 3.2.1.0  │ CVE-1234-5678  │ HIGH     │\n└──────────────┴───────────────┴──────────┴────────────────┴──────────┘\n"
         self.mock_file.seek(0)  # reset file position
         self.assertIn(expected_output, self.mock_file.read())
@@ -128,21 +132,21 @@ class TestOutputEngine(unittest.TestCase):
         if "Output stored at" in cm.output[0]:
             contains_msg = True
 
-        self.assertEqual(contains_filename, True)
-        self.assertEqual(contains_msg, True)
-
         # reset everything back
         os.remove(filename)
         self.output_engine.filename = ""
+
+        self.assertEqual(contains_filename, True)
+        self.assertEqual(contains_msg, True)
 
     def test_output_file_filename_already_exists(self):
         """Tests output_file when filename already exist"""
 
         # update the filename in output_engine
-        self.output_engine.filename = "testfile"
+        self.output_engine.filename = "testfile.csv"
 
         # create a file with the same name as output_engine.filename
-        with open("testfile", "w") as f:
+        with open("testfile.csv", "w") as f:
             f.write("testing")
 
         logger = logging.getLogger()
@@ -155,7 +159,7 @@ class TestOutputEngine(unittest.TestCase):
         msg_generate_filename = (
             "Generating a new filename with Default Naming Convention"
         )
-        msg_failed_to_write = "Failed to write at 'testfile'. File already exists"
+        msg_failed_to_write = "Failed to write at 'testfile.csv'. File already exists"
 
         # flags for logs
         contains_fail2write = False
@@ -169,7 +173,7 @@ class TestOutputEngine(unittest.TestCase):
                 contains_fail2write = True
 
         # remove the generated files and reset updated variables
-        os.remove("testfile")
+        os.remove("testfile.csv")
         os.remove(self.output_engine.filename)
         self.output_engine.filename = ""
